separate classes into dedicated files for ext models
