+++ src/Altinn.Correspondence.API/Models/InitializeCorrespondencesResponseExt.cs
@@ -1,45 +1,21 @@
 using System.Text.Json.Serialization;
-using Altinn.Correspondence.API.Models.Enums;
 
 namespace Altinn.Correspondence.API.Models;
 
+/// <summary>
+/// Response object for initializing correspondences. Contains information about the created correspondences and their attachments.
+/// </summary>
 public class InitializeCorrespondencesResponseExt
 {
+    /// <summary>
+    /// The initialized correspondences
+    /// </summary>
     [JsonPropertyName("correspondences")]
     public List<InitializedCorrespondencesExt> Correspondences { get; set; }
+
+    /// <summary>
+    /// The IDs of the attachments that were included with the correspondences
+    /// </summary>
     [JsonPropertyName("attachmentIds")]
     public List<Guid> AttachmentIds { get; set; }
-}
-
-
-public class InitializedCorrespondencesExt
-{
-    [JsonPropertyName("correspondenceId")]
-    public Guid CorrespondenceId { get; set; }
-
-    [JsonPropertyName("status")]
-    public CorrespondenceStatusExt Status { get; set; }
-
-    [JsonPropertyName("recipient")]
-    public required string Recipient { get; set; }
-
-    [JsonPropertyName("notifications")]
-    public List<InitializedCorrespondencesNotificationsExt>? Notifications { get; set; }
-}
-public class InitializedCorrespondencesNotificationsExt
-{
-    [JsonPropertyName("orderId")]
-    public Guid? OrderId { get; set; }
-
-    [JsonPropertyName("isReminder")]
-    public bool? IsReminder { get; set; }
-
-    [JsonPropertyName("status")]
-    public InitializedNotificationStatusExt Status { get; set; }
-}
-public enum InitializedNotificationStatusExt
-{
-    Success,
-    MissingContact,
-    Failure,
 }
